Add minimum stock to forester and farmer
These are crafters so have request system conflicts, but it also makes
sense for them to have minimum stocks (for tools, and potentially for
seeds/saplings).

The other crafters either would only have reason to request materials,
which can be warehouse-stocked, or fuel, which is more complicated and
should usually be stocked anyway.